Inherit from LWP::RobotUA.
I had it commented-out because it was sleeping when I ran it to hit the first
site. This was annoying. It turns out it does this because it counts the
initial request for robots.txt as a site visit, and will wait before visiting
the site again.

So work around it by overriding `host_wait()` so that it returns false for the
first *and* second request to a URL. And now I've re-established the delay and
inheritance from LWP::RobotUA.

Later I might set `use_sleep => 0`. Will have to see how I end up managing
parallel requests.