Generate our own IDs.
Using a combination of the site URL and the entry URL. This will allow us to
have two blogs with the same URL, as can happen when one blog links to an
entry in another blog, and we happen to manage both. It also makes the IDs
consistent, which is kinda nice.