Parse audio and video from content.
Also be sure to eclude any enclosures or embeds that are not audio, video, or
images.

Ugly bug in XML::LibXML where `recover => 2` does not prevent it from issuing
warnings for unrecognized tags (audio and video in this case). I'm disabling
warnings to prevent the warnings, and it *still* emits crap to (I assume)
STDERR. Will have to submit a bug report about that. Hopefully libxml2 will be
updated to recognize those tags soon anyway.