Follow redirects for unknown types.
If looking for media and the URL doesn't seem to have an expected extension,
send a request to the URL and see what its headers say it is. This is
especially handy for redirects.

But need to add code to use the redirected URL as the enclosed URL, rather
than the redirecting URL. That's next.