Fix bug in UA::Robot.
The code was forcing LWP::UserAgent::WithCache to inherit from LWP::RobotUA.
That was fine for its code, but App::FeedScene::UA also needed to use
WithCache, but *without* the robot stuff. The fix is an even uglier hack, as
it just imports all of the LWP::RobotUA methods directly into the UA::Robot
module, the `new()` method temporarily forces WithCache to inherit from
RobotUA so that we get the configuration right, and the `host_wait()` method
must dispatch directly to RobotUA's `host_wait()` method. It's all ugly, nasty
hackery, but at least its encapsulated.

The upshot is that I'm able to use App::FeedScene::UA directly again in
EntryUpdater. Should have known that it was my bug that was slowing things
down. I'm glad to have it fixed.