Add `fs:sources` element output.
This is obviously not an Atom-compliant element, so I encapsulate it into a
namespace in case it gets added at some time in the future. Also, I don't
output it if the strict Atom generation is requested.

While at it, use kineticode.com for the domain name for now, so that I can
actually put things to use using the Kineticode server until we have a new
domain.