Add HTML tidying.
Right now it falls back on cleaned up HTML if XML::LibXML dies, but I think I
might move it to always tidy. We'll see.

This came about because I found a nasty example of a summary generated from
content with an anchor tag. I thought it was because the HTML was invalid, so
I added the HTML cleanup. It turns out that the HTML actually wasn't so bad
(though it would fail to parse), but that I had another bug in `_clean_html()`:
it wouldn't move back up to the parent's sibling if it was throwing out a
tag, it would only go into its siblings. Oops.

While at it, I've added code to ignore elements that have no child nodes or
attributes, as empty tags are useless. I've also banned `br` tags.

Oh, and I moved all the HTML parsing into App::FeedScene::Parser. Gotta look
into moving the ZapCP1252 stuff into there, too.

Also fixed the names of some of the existing summary-testing tests.