Rmove HTML::Tidy dependency.
It's still in there, but commented-out for now. It turns out that XML::LibXML
seems able to do the same sort of cleanup. Happy to eliminate the dependency

While at it, even though it's commented out, I separated the inline from the
block HTML 5 elements in the creatin of the HTML::Tidy object. That will be
handy should I need to go back to using HTML::Tidy. Hoping I don't have to.