Handle CP1252 gremlins.
The key was to update Encode::ZapCP1252 so that it works as well on decoded
strings as it does on encoded strings. That's 0.20, now on CPAN (and GitHub).

Before i figured that out, I went down a rabbit hole of trying to detect the
original encoding of each XML file. That's what the `parse_encoding()` method
in Parser tries to figure out. I may remove that, though; gotta figure out how
to deal with completely broken characters.