Fix broken characters from Pipes feed.
The issue is that, even though the characters were broken in the Pipes feed,
they were still valid UTF-8. But by running `zap_cp1252` on them, I was
breaking those characters. I got XML::LibXML to choke on it, and discovered
that it stopped choking when I stopped zapping. Furthermore, CP1252 characters
that I'd been zapping were suddenly coming out as nice UTF-8, becaue I'm
fixing them elsewhere now. So this is overall an improvement.

While at it, I changed `parse_feed()` to take an HTTP::Response object as
its argument, rather than XML. This is so that I can next test with a truly
broken feed with invalid bytes in it and figure out how to recover from
the parse and try to fix it. That will come next.