Add support for some more elements.
Found an exmple in the wild where content was missing because it was inside a
`<blockquote>` element. So I added that, along with a test, and added the code
to convert it to a `<div>`.

Reviewing the list of HTML 5 elements, I also added `<aside>` and `<footer>`.
Kind of on the fence about `<hr>` and `<br>`. I might have to allow for them
later.