Ignore doubleclick.net images.
We have at least one feed that's using doubleclick.net images as the first
images in the content. This doesn't look so great in the feed -- even when
it's using an invisible single-pixel gif (as I saw in the example). So just
exclude any images from doubleclick.net.

With the change to the tests, the generator tests broke. This is because all
the enclosure entries had the same publish date. So make sure they have unique
publish dates.