Strip out undesired characters.
One of our feeds was making XML::LibXML choke. The reason? It has a control
character in it, '\x03' (same as `chr 3`, lookes like "^C" in Emacs). Thanks
to help from hobbs on IRC, who pointed me at
http://www.w3.org/TR/xml11/#charsets, I was able to catch this condition and
make the effort to simply strip out verboten characters (they're valid UTF-8,
which is why Encode doesn't remove them). The code now also dies if that has
been tried, so we avoid the infinite loop there (which was killing the server
last week).

I'm told that it may be that control characters are allowed in some parts of
XML (perhaps CDATA blocks), but I think it's safest just to strip them out
altogether, since it'd be more of a bitch to write a chunk parser and only
strip them out of where they don't belong. I'm pretty sure they are
unintentionally used the vast majority of the time, and highly unnlikely to be
useful in our app, anyway. So just kill them all.