Handle non-encoded XML characters.
Logopond had a title with a raw "&" in it, and that made XML::LibXML choke,
because that character should always be encoded in XML.

So use the error context from the XML::LibXML error to find the bogus character
in the XML, entitize it, and try again.

While at it, make sure that `t/parser.t` deletes its cache after it runs. This
was causing some headaches when I modified the XML and my changes didn't seem
to do anything -- subsequent test runs were using the cached copy.