Track request failures.
Add the `fail_count` column to the `feeds` table and use it to track request
failures. Only output request failures when we have more than a certain
threshold *and* then only every x number of failures. Once there's a success
(or a not modified result), reset the count to 0.

This should bring down the number of emails I get every day, and allow me to
track only those that are consistenly poor responders, which generally will
mean that they're gone. Maybe.