Find best Flickr photos.
Add `_audit_enclosure()` method. This method will handle any transformations
of enclosure URLs. Right now, it just returns the URL unless it's a Flickr
Photo URL, in which case it uses the Flickr API to find the best image it
can for the iPad.

Need to add some code to cache these calls.