Ignore non-XML responses.
This is mainly to avoid lots of errors from Tumblr, which stupidly returns an
HTML page and no bloody content type header when a feed is down for service.
So we now check the `content_is_xml` method and, if it returns false, log it
as a failure and skip to the next one. This will eliminate those Data::Feed
parse failures.

This has been bugging me for a while; I'm glad to finally have a solution.