Try harder to determine if a feed is a feed.
One of our feeds, http://www.fastcodesign.com/rss.xml, as a content-type
header that says it's HTML, even though the content is actually RSS. So if a
feed says it's not a feed in the content type, we check the body of the thing,
too, in order to make sure. I'm just using Data::Feed's heuristics for this.