Use --app instead of --file for distribution.
This makes it more like fs_gen and the others. It will find the file to
distribut itself. Also, allow the bucket to default to "feedscene".
Use --app instead of --file for distribution.

This makes it more like fs_gen and the others. It will find the file to
distribut itself. Also, allow the bucket to default to "feedscene".