Fix over-aggresive cutting of dupes.
Noticed lately how, in DesignScene, you could see an image in one sync and
have it replaced by a similar, but different image in the next sync? Or even
have it disappear? This has been handy for tracking down and fixing bugs in
how portals update, because that sort of thing *can* happen. But the fact that
it was happening all the time was, of course a bug.

The issue is that I was being too aggressive in disallowing dupes, to the
extent that a single entry couldn't use the same image from one sync to
another. In fact, a Flickr user couldn't even have an image display for more
than one sync, although it might reapear in a later sync. Oops.

So fixed by exlucding the current entry's ID when searching for dupes. This
means that an entry can continue using the same image. Also Exclude the
current entry ID when looking for duplicate Flickr usernames. Same issue. This
should make the appearence of images in DS much more stable and predicable.