Fetch enclosure only once per entry.
Can always back this out if we find much stuff drops off, or can add
validation for the fetched URL, but since a changing image should be the
exception, rather than the rule, this should do, and will allow the script to
run as fast as possible, becaus it won't have to fetch very many new media
documents from the Internet.

While at it, refactor `t/entry_updater.t` so that the mapping of media types
to URLs is much more reasonable. There's really only one that's not text/html,
so creat a mapping table that'll be easier to update. I could have saved
myself an hour's work today and 2-3 yesterday if I'd done this before.