Ignore unfavorited Flickr Pool photos.
This will hopefully prevent the Flickr pools from overwhelming everything
else. We now require that photos in a pool be favorited before we show them.
The code is written such that it'd be easy to require a minimum number of
favorites, too.

A pool, right now, is simply identified as any feed with the host name
api.flickr.com. All of our pool URLs currently are on this host, and that's
unlikely to change. We can always tweak our definition of a pool, as well
(currently, all but one of our pool URLs calls the `groups_pool.gne` method;
the exception is a URL that calls `photos_public.gne`).