Switch to XML::Liberal.
By using XML::Liberal, we get some of the same error-handling and correction
that I was already doing, plus a whole bunch more that we haven't yet run
into. One we got a while ago, a bogus character in a file, was not caught by
the recovery stuff I'd written so far, but XML::Liberal handles it great.

I also went through the other XML files I had with errors. All of them parse
fine with XML::Liberal. So I think we're good to go. Closes #10. Closes #11.